Fix migration start failurein Icehouse
Due to some VNX side sync issue, a migration session will fail
when it uses a destination LUN id which also involves in a
previous finished session. the errror message is:
Error: migrate -start command failed.
The destination LUN is not available for migration.

This patch will retry to start the session when above error occurs.